backend/requirements.txt
aiofiles==24.1.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
asyncpg==0.30.0
certifi==2024.8.30
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
dnspython==2.7.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.4
fastapi-cli==0.0.5
fonttools==4.54.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
imageio==2.36.0
ipydatawidgets==4.3.5
ipython==8.29.0
ipywidgets==8.1.5
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
lazy_loader==0.4
lxml==5.3.0
Mako==1.3.6
mapbox_earcut==1.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pythreejs==2.4.2
PyYAML==6.0.2
rich==13.9.4
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.2
svg.path @ git+https://github.com/regebro/svg.path.git@867a09b8b2bffd1b5a0b066bc39f993f21e3e708
svgpathtools==1.6.1
svgwrite==1.4.3
tifffile==2024.9.20
traitlets==5.14.3
traittypes==0.2.1
trimesh==4.5.2
typer==0.13.0
typing_extensions==4.12.2
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
widgetsnbextension==4.0.13
